Path Param - Rest routes

+++ src/main/java/com/sk/explorecamel/model/NotifcationModel.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class NotifcationModel {
 
-	private Integer feedId;
+	private Long feedId;
 	private String headline;
 	private String content;
 	private Date lastEditedDate;
